backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=5.0.0
biopython>=1.80
gunicorn>=21.2.0
python-dotenv>=1.0.0
django-celery-results>=2.5.0
djangorestframework-simplejwt>=5.3.0
primer3-py>=2.0.0
abnumber>=0.3.0
anarci>=2.0.0
